Replace CopyOnto by Push (#28)
* Adding several implementations that were missing.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Replace CopyOnto by Push

Swap the parameters for CopyOnto. Previously, we'd implement CopyOnto
for a specific type, specifying the region as a parameter to CopyOnto. This
has the undesired side-effect that downstream crates cannot implement the
trait for their own regions due to Rust's orphan rules. Switching the
parameters around (Push<T> for Region) moves a local type into the T0
position, which is compatible with the orphan rules.
---------

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>